requirements.txt
requests~=2.25.1
beautifulsoup4~=4.9.3
tqdm~=4.60.0
psycopg2-binary~=2.8.6
